requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
gspread
google-auth
xlsxwriter
